refactor(go): keep gemini code in googlegenai
