+++ go/plugins/googlegenai/models.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 
 	"github.com/firebase/genkit/go/ai"
-	"github.com/firebase/genkit/go/plugins/internal/gemini"
 	"google.golang.org/genai"
 )
 
@@ -57,7 +56,7 @@ var (
 				"gemini-1.5-flash-001",
 				"gemini-1.5-flash-002",
 			},
-			Supports: &gemini.Multimodal,
+			Supports: &Multimodal,
 		},
 		gemini15Pro: {
 			Label: "Gemini 1.5 Pro",
@@ -66,7 +65,7 @@ var (
 				"gemini-1.5-pro-001",
 				"gemini-1.5-pro-002",
 			},
-			Supports: &gemini.Multimodal,
+			Supports: &Multimodal,
 		},
 		gemini15Flash8b: {
 			Label: "Gemini 1.5 Flash 8B",
@@ -74,41 +73,41 @@ var (
 				"gemini-1.5-flash-8b-latest",
 				"gemini-1.5-flash-8b-001",
 			},
-			Supports: &gemini.Multimodal,
+			Supports: &Multimodal,
 		},
 		gemini20Flash: {
 			Label: "Gemini 2.0 Flash",
 			Versions: []string{
 				"gemini-2.0-flash-001",
 			},
-			Supports: &gemini.Multimodal,
+			Supports: &Multimodal,
 		},
 		gemini20FlashLite: {
 			Label: "Gemini 2.0 Flash Lite",
 			Versions: []string{
 				"gemini-2.0-flash-lite-001",
 			},
-			Supports: &gemini.Multimodal,
+			Supports: &Multimodal,
 		},
 		gemini20FlashLitePrev: {
 			Label:    "Gemini 2.0 Flash Lite Preview 02-05",
 			Versions: []string{},
-			Supports: &gemini.Multimodal,
+			Supports: &Multimodal,
 		},
 		gemini20ProExp0205: {
 			Label:    "Gemini 2.0 Pro Exp 02-05",
 			Versions: []string{},
-			Supports: &gemini.Multimodal,
+			Supports: &Multimodal,
 		},
 		gemini20FlashThinkingExp0121: {
 			Label:    "Gemini 2.0 Flash Thinking Exp 01-21",
 			Versions: []string{},
-			Supports: &gemini.Multimodal,
+			Supports: &Multimodal,
 		},
 		gemini25ProExp0305: {
 			Label:    "Gemini 2.5 Pro Exp 03-25",
 			Versions: []string{},
-			Supports: &gemini.Multimodal,
+			Supports: &Multimodal,
 		},
 	}
